+++ AddOns/XIV_Databar_Continued/Core/changelog/2.8.lua
@@ -1,0 +1,48 @@
+local AddOnName, XIVBar = ...;
+
+XIVBar.Changelog[2800] = {
+    version_string = "2.8",
+    release_date = "2024/05/16",
+    important = {
+        ["zhCN"] = {},
+        ["zhTW"] = {
+            "[正式版] 更新 TOC 的遊戲版本編號 10.2.7。",
+            "[經典版] 更新探索賽季 TOC 的遊戲版本編號 1.15.2。",
+        },
+        ["enUS"] = {
+            "[Retail] TOC update for Retail 10.2.7.",
+            "[Classic] TOC update for Classic SOD 1.15.2.",
+        },
+        ["frFR"] = {
+            "[Retail] Mise à jour du TOC pour la version 10.2.7 de Retail.",
+            "[Classic] Mise à jour du TOC pour la version 1.15.2 de Classic - Saison de la découverte.",
+        },
+        ["koKR"] = {},
+        ["ruRU"] = {}
+    },
+    new = {
+        ["zhCN"] = {},
+        ["zhTW"] = {
+            "[浩劫與重生經典版] 初次支援 4.4.0。",
+            "[浩劫與重生經典版] 已知問題 : 天賦切換有問題，已計劃會盡快修復。",
+        },
+        ["enUS"] = {
+            "[Cata Classic] Initial support for Cata Classic 4.4.0.",
+            "[Cata Classic] Known bug : The talent switcher is actually bugged, a fix is planned ASAP.",
+        },
+        ["frFR"] = {
+            "[Cata Classic] Mise en compatibilité initiale pour Cata Classic 4.4.0",
+            "[Cata Classic] Bug connu : Le sélecteur de talents est actuellement non fonctionnel, un correctif sortira dès que possible.",
+        },
+        ["koKR"] = {},
+        ["ruRU"] = {}
+    },
+    improvment = {
+        ["zhCN"] = {},
+        ["zhTW"] = {},
+        ["enUS"] = {},
+        ["frFR"] = {},
+        ["koKR"] = {},
+        ["ruRU"] = {}
+    }
+}