+++ AddOns/XIV_Databar_Continued/CHANGELOG.md
@@ -1,9 +1,18 @@
 # XIV_Databar Continued
 
-## [2.7.3](https://github.com/ZelionGG/XIV_Databar-Continued/tree/v2.7.3) (2024-03-22)
+## [2.8](https://github.com/ZelionGG/XIV_Databar-Continued/tree/v2.8) (2024-05-16)
 
-[Full Changelog](https://github.com/ZelionGG/XIV_Databar-Continued/compare/v2.7.2...v2.7.3) [Previous Releases](https://github.com/ZelionGG/XIV_Databar-Continued/releases)
+[Full Changelog](https://github.com/ZelionGG/XIV_Databar-Continued/compare/v2.7.3...v2.8) [Previous Releases](https://github.com/ZelionGG/XIV_Databar-Continued/releases)
 
 ### _Retail :_
 
-- TOC update for Retail 10.2.6.
+- TOC update for Retail 10.2.7.
+
+### _Classic (SOD) :_
+
+- TOC update for Classic SOD 1.15.2.
+
+### _Cata Classic :_
+
+- Initial support for Cata Classic 4.4.0.
+- Known bug : The talent switcher is actually bugged, a fix is planned ASAP.